LGA-2590 add helper functions to common_steps
